pyproject.toml
[project]
name = "ppo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gym-minigrid>=1.2.2",
    "gymnasium>=1.2.1",
    "matplotlib>=3.10.7",
    "minigrid>=3.0.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "torch>=2.9.0",
]
